Each variable within MPGTextField.m file declared as private.
Array is always filtered when the filtered array is required. This is
an overhead when the data has 100k of words. It’s too slow and blocks
UI thread.

This should be like filter the array and store in controller and used
that private filtered array when required. Do not need to filter an
filter again.